+++ lib/api/api_client.dart
@@ -16,7 +16,4 @@ class ApiClient extends Dio {
   Map<String, String> get headers => options.headers;
 
   set headers(Map<String, String> newHeaders) => options.headers = newHeaders;
-
-  @deprecated
-  void setDefaultHeader(Map<String, String> headers) => this.headers = headers;
 }