+++ lib/utils/validator/rules/any_rule.dart
@@ -2,19 +2,17 @@ import 'package:flrx/utils/validator/rule.dart';
 
 class AnyRule<T> extends Rule<T> {
   final List<Rule<T>> _ruleList;
-  final String validationMessage;
 
-  AnyRule(this._ruleList, {this.validationMessage});
+  AnyRule(this._ruleList, {String validationMessage})
+      : super(validationMessage);
 
   @override
-  String validate(String entityName, T value) {
+  String onValidate(String entityName, T value) {
     String ruleValidationMessage;
     _ruleList.reversed.any((Rule<T> rule) {
       ruleValidationMessage = rule.validate(entityName, value);
       return ruleValidationMessage == null;
     });
-    return ruleValidationMessage == null
-        ? ruleValidationMessage
-        : (validationMessage ?? ruleValidationMessage);
+    return ruleValidationMessage;
   }
 }