+++ lib/utils/validator/rules/required_rule.dart
@@ -1,8 +1,10 @@
 import 'package:flrx/utils/validator/rule.dart';
 
 class RequiredRule extends Rule<String> {
+  RequiredRule({String validationMessage}) : super(validationMessage);
+
   @override
-  String validate(String entityName, String value) {
+  String onValidate(String entityName, String value) {
     if (value.isEmpty) {
       return '$entityName should not be empty!';
     }